whole calculator nearly done
added a config file with the map of the operators and their strength.
also added a printer module that has all the print functions in it.
added the digitSum operator (#) and implemented it.
tried cleaning up the code again to make it more readable.
added a function that shows the user exactly where an error occurred.
there might be a few edge cases that are not implemented yet and will be implemented if I find any.